pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "meetcap"
version = "0.1.0"
description = "Zoom meeting transcription and summarization utility"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "sounddevice>=0.4.6",
    "numpy>=1.24.0",
    "watchdog>=3.0.0",
    "anthropic>=0.18.0",
    "openai-whisper>=20231117",
    "pyyaml>=6.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
meetcap = "meetcap.cli:main"
meetcap-watch = "meetcap.watcher:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
